Tweak README files and build script slightly based on sudo-67.1/sudo-72
The sudo-67.1 (10.8.5) and sudo-72 (10.9.0) sources have been released
so tweak the README files to reflect that and based on the configure
options listed for sudo-72 (10.9.0) re-enable the three configure
options (--disable-setreuid --with-env-editor --with-password-timeout=0)
when building for 10.7.x/10.8.x that were apparently mistakenly omitted
from those builds as released with OS X.

The --with-env-editor and --with-password-timeout=0 options have minor
effect and were always present prior to 10.7.0 (and are again in 10.9.0).

The --disable-setreuid option indicates that OS X's setreuid is not
compatible with sudo.  This option was always set prior to 10.7.0.  It's
rather inconceivable that OS X's setreuid functionality was fixed for
both 10.7.0 and 10.8.0 only to be broken again in 10.9.0.  This just
gives more weight to the notion that omitting those three options from
the 10.7.x/10.8.x sudo OS X builds was a mistake.  Therefore those
three options are now set when building on 10.7.x/10.8.x as well.